Change CommandMachine get state from out files
It used to keep track of its state by touching a file that
corresponds to each flag. Instead it's now using the presence of
an output file to indicate whether the flag for that state is set
or not. This is similar to how make works. This will allow us to
easily restart a failed process.