Write all results to temp files first
Changed it so that for each step we write the results to a temp
file and then rename the temp file to the final location. This
means that a file that we expect to exist after a step is run is
either not there at all or is complete. This should help keep the
state of the pipeline clean, in that we can kill a job at any point
and it won't leave any partially finished files. Well, it might
leave some partially finished *temporary* files, but those at least
won't be confused with the *real* files.