Working on #26, clean up after aligning
I changed it so "rum_runner align" automatically runs the
RUM::Action::Clean action after a successful run. In order to make
this work I had to redo some of the task dependencies. For example,
after postprocessing is done, we create the final mapping_stats.txt
file, based on u_footprint.txt and nu_footprint.txt, and then
we remove those files. That made it so we can't run "rum runner align"
again without rerunning the tasks that generate *footprint.txt. I just
moved the stat calculation out of RUM::Action::Align and into its
own step of the workflow, so now it is idempotent.