Fix #46: infinite loop in parse_blat_out.pl
parse_blat_out.pl attempts to skip over the header lines in the
blat output. It's basically looking for the first nonempty line
that doesn't look like a header line. If blat didn't find any
mappings, there will be an EOF right after the header, which will
cause this loop to continue forever. Fixed by checking whether
the line is defined, which is good practice anyway.