Working on #47 (bug in mapping_stats.txt)
mapping_stats.pl requires its input files to be sorted on id.
The workflow in RUM 2.0 was running it on the RUM_* files that
are sorted by location. I changed it so that we instead run it
on the files sorted by id. These files only existed during the
processing phase, not in postprocessing, and they are never
merged together. So I modified count_reads_mapped.pl so that
it accepts a list of RUM_Unique files and a list of RUM_NU files,
rather than concatenating all those files together into a
separate file.