Fix #50: can't run "rum_runner help" from base dir
"rum_runner help" uses Pod::Usage, which by default runs perldoc
in order to generate the usage message. It turns out that if you
run perldoc as root, it turns on taint-checking, which prevents
perl from doing certain "unsafe" things. Also, if you run perldoc
in a directory that contains a file called Makefile.PL, it will
attempt to do some things that it considers "unsafe". So the
bottom line is that you can't run perldoc (or any program that
uses perldoc, such as "rum_runner help") as root in a directory
that contains Makefile.PL.

I just applied a quick fix that makes it not use perldoc if it's
running as root in a directory that contains Makefile.PL. The help
output will be a little uglier in that case, but it's better than
nothing.