Die if we can't read from an input file (#55)
If a user runs rum_runner and has a stray argument on the command
line (for example if they accidentally add an argument to a switch
that doesn't accept one, like "--limit-bowtie-no 50"), we will
interpret that stray argument as an input read file. This change
is just to make it so that we fail fast if we see an input read
file that we can't read.