Fix #56: don't restart preproc if not needed
We were determining whether or not we need to run preprocessing
based on whether *all* of the chunks have finished at least
one step of their workflow. This means that if one chunk never
starts the workflow, we will always rerun preprocessing in order
to rerun that chunk.

This is a little more strict than we need to be. Since we don't
start any of the chunks until all of preprocessing is done, we can
simply check to see if *any* of the chunks have finished at least
one step of the workflow.

A better solution would be to model preprocessing as a
RUM::Workflow, like the per-chunk processing and the postprocessing
phases are modeled. However, that would require some serious
refactoring of the preprocessing phase, and it doesn't seem worth
it right now.