Update test file since index changed
I have been getting failures on the test for
make_TU_and_TNU.pl for a little while. It turns out that when
Greg remade the index for Arabidopsis to fix an issue, some of the
gene locations changed. This test uses output from Bowtie (Y.1)
and parses it using make_TU_and_TNU.pl. The Y.1 file was based on
the old indexes, and make_TU_and_TNU.pl was accessing the new
indexes. Updating the Y.1 file fixed it.