Fix #70: allow read files to be moved
It used to be that rum_runner would check to make sure that the
raw read files were present every time it runs. A few users have
mentioned that they typically zip up the input files once the job
starts, so it would be nice if rum_runner didn't require the read
files to be present after it's done with them.

Now we only check that the original raw files are present if the
"reads.fa" file is *not* in the output directory.