Fix #100: qsub error
Apparently SGE jobs can't start with a digit. We name the scripts
we submit based on the --name the user provides to rum_runner.
Before we were just calling the scripts "${name}_{phase}.sh". So
if $name started with a digit, it would fail. I just changed it so
we prepend "rum_" to the name, which fixes it.

Also changed it so we log the STDERR output from qsub, to make
qsub errors easier to debug in the future.