Fix #102: error creating index files.
There should be a step in the index creation workflow that
modifies the headers in the genome fasta file so that they
only contain the chromosome name. It looks like I removed
that step some time ago, probably accidentally. This was
causing some downstream steps to produce empty files. I
added the step back, and it seems to fix everything.