Add version number to job_settings file (#122)
The format of the .rum/job_settings file changed between
v2.0.2_01 and v2.0.2_02. Trying to run RUM > 2.0.2_01 on a
job that was initialized with RUM <= 2.0.2_01 would fail
in a nasty way.

I changed it so that every time the job_settings file is saved,
we record the version of RUM that saved it. This will allow us
to be backwards compatible with certain version in the future.
For now, since RUM 2.x is not in very widespread use, I am
simply making it so that running RUM >= 2.0.2_04 on a job
that doesn't have the version number recorded in it will fail.
This should at least make it clear to the user what is going
on.