Make 'status' show preprocessing status (fix #91)
Now 'rum_runner status' will show a line of Xs for the
preprocessing phase, one X for each chunk whose input
has been split out from the main input file. This is
maybe not the best interface for it, but it's at least
consistent with the rest of the interface, and wasn't too
much work to add.