Move job settings file out of .rum (#117, #128)
A few people have noticed that if you do "rm $output_dir/*",
the .rum directory does not get removed. I moved all of the
files out of the .rum directory and put them in the main
output directory for the job. If you run thew new version of
RUM on an older job that still has the .rum directory, it
should still work, but just print a warning about the old
configuration file.