Close #133: fail gracefully if index not found
This should improve the way we handle a case where a user runs a
command on an old job after the index has moved. If you run status,
kill, clean, or stop, it should now just print a warning that the
index has moved and proceed with the action. If you run align or
resume, which actually need the index, it will die with a more
descriptive error message.