Make rum2sam explicity check for corrupt input
We've seen a few cases where sort_RUM_by_id.pl produces files
that have more or less fields than they should on a few lines.
This can cause rum2sam to enter an infinite loop. I'm adding a
check every time we read a line in to make sure that it at least
has the right number of fields.