Fix duplicate mapping (#158)
This was a weird edge case where merge_GU_and_TU would (rarely)
add a duplicate copy of a mapping. It looks like it has to do
with how the script handles blocks of records at a time in order
to avoid using too much memory. It was inadvertently retaining
the last record from one block when starting another block.